requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9

langgraph>=0.1.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langsmith>=0.1.0

openai>=1.30.0

qdrant-client>=1.9.0

pandas>=2.2.0
openpyxl>=3.1.0

python-dotenv>=1.0.0
pydantic>=2.7.0

slack-sdk>=3.27.0
jira>=3.8.0

langfuse>=2.20.0

pytest>=8.2.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
